pyproject.toml
[tool.poetry]
name = "dagsheets"
version = "0.1.0"
description = ""
authors = ["ayslanleal <ayslanleal15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.0"
pandas = "^1.4.2"
google-api-python-client = "^2.51.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
